Update specification.html
Update examples for AG4 and AG5 so that AG4 includes a case where a schema.org property must be used because the Agent to Agent Relationship Roles vocabulary does not have "alumni of" concept, and therefore a dcat:qualifiedRelationship pattern is not used there. AG5 stipulates the dcat pattern and uses AARR vocab.